fixed the not inserting problem: I was forgetting to call .await in the tokio related functions I moved away from main.rs
