+++ .tool-versions
@@ -1,0 +1,1 @@
+rust 1.72.0
